@@ -1,5 +1,5 @@
 {
-  "name": "szunisoft/laravel-mnb",
+  "name": "csaba215/laravel-mnb",
   "license": "MIT",
   "description": "Laravel MNB (Magyar Nemzeti Bank)",
   "keywords": [
@@ -20,26 +20,26 @@
     }
   ],
   "require": {
-    "php": "^7.1.3",
-    "szunisoft/mnb": "^1.0.1",
-    "illuminate/support": "5.* | 6.* | 7.*",
-    "illuminate/cache" : "5.* | 6.* | 7.*"
+    "php": "^8.1",
+    "illuminate/support": "10.* | 11.* | 12.*",
+    "illuminate/cache" : "10.* | 11.* | 12.*",
+    "ext-soap": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.5",
-    "mockery/mockery": "1.3.*",
-    "orchestra/testbench": "^5.1.0"
+    "phpunit/phpunit": "^11.5.3|^12.0.1",
+    "mockery/mockery": "^1.6.10",
+    "orchestra/testbench": "^10.2.1"
   },
   "autoload": {
     "files": [
     ],
     "psr-4": {
-      "SzuniSoft\\Mnb\\Laravel\\": "src/"
+      "Csaba215\\Mnb\\Laravel\\": "src/"
     }
   },
   "autoload-dev": {
     "psr-4": {
-      "SzuniSoft\\Mnb\\Laravel\\Tests\\": "tests/"
+      "Csaba215\\Mnb\\Laravel\\Tests\\": "tests/"
     }
   },
   "config": {
@@ -53,7 +53,7 @@
   "extra": {
     "laravel": {
       "providers": [
-        "SzuniSoft\\Mnb\\Laravel\\MnbServiceProvider"
+        "Csaba215\\Mnb\\Laravel\\MnbServiceProvider"
       ]
     }
   }
