@@ -1,125 +1,92 @@
 <?php
 
 
-namespace SzuniSoft\Mnb\Laravel;
-
+namespace Csaba215\Mnb\Laravel;
 
 use Illuminate\Contracts\Cache\Repository;
-use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Cache;
+use SoapClient;
+use SoapFault;
 
-/**
- * Class Client
- * @package SzuniSoft\Mnb\Laravel
- */
-class Client {
+class Client
+{
+    protected SoapClient $client;
 
-    /**
-     * @var \SzuniSoft\Mnb\Client
-     */
-    protected $client;
+    protected Repository $cache;
 
     /**
-     * @var Repository
+     * @throws SoapFault
      */
-    protected $cache;
-
-    /**
-     * @var int
-     */
-    protected $timeout;
-
-    /**
-     * Client constructor.
-     * @param \SzuniSoft\Mnb\Client $client
-     * @param \Illuminate\Cache\Repository $cache
-     * @param $timeout
-     */
-    public function __construct(
-        \SzuniSoft\Mnb\Client $client,
-        \Illuminate\Cache\Repository $cache,
-        $timeout
-    )
+    public function __construct()
     {
-        $this->client = $client;
-        $this->cache = $cache;
-        $this->timeout = $timeout;
-    }
-
-    /**
-     * @param $date
-     */
-    protected function normalizeDate(&$date)
-    {
-        $date = Carbon::createFromFormat('Y-m-d', $date);
+        $this->client = new SoapClient(config('mnb-exchange.wsdl'));
+        $this->cache = Cache::store(config('mnb-exchange.cache.store'));
     }
 
     /**
      * @return string[]
      */
-    public function currencies()
+    public function currencies(): array
     {
-        return $this->cache->remember('mnb.currencies', $this->timeout, function () {
-            return $this->client->currencies();
-        });
+        return $this->cache->remember(config('mnb-exchange.cache.key').'.currencies', config('mnb-exchange.cache.minutes'),
+            fn () => (array) simplexml_load_string($this->client->GetCurrencies()->GetCurrenciesResult)->Currencies->Curr
+        );
     }
 
     /**
-     * @param $code
-     * @return bool
+     * Determines if the currency is supported or not.
      */
-    public function hasCurrency($code)
+    public function hasCurrency(string $currency): bool
     {
-        return in_array($code, $this->currencies());
+        return in_array($currency, $this->currencies());
     }
 
     /**
      * Will return with the cached exchange rate
-     *
-     * @param $code
-     * @param null $date
-     * @return null|\SzuniSoft\Mnb\Model\Currency
      */
-    public function exchangeRate($code, &$date = null)
+    public function exchangeRate(string $code, $date = null): array
     {
-        return $this->cache->remember("mnb.currencies.rate.$code", $this->timeout, function () use (&$code, &$date) {
-            return $this->currentExchangeRate($code, $date);
-        });
-    }
+        if ($date === null) {
+            $date = $this->lastOpeningDate();
+        }
 
-    /**
-     * @param null $date
-     * @return array|\SzuniSoft\Mnb\Model\Currency[]
-     */
-    public function exchangeRates(&$date = null)
-    {
-        return $this->cache->remember("mnb.currencies.rate", $this->timeout, function () use (&$date) {
-            return $this->currentExchangeRates($date);
-        });
+        return $this->cache->remember(config('mnb-exchange.cache.key').".currencies.rate.$code.$date", config('mnb-exchange.cache.minutes'),
+            function () use ($date, $code) {
+                $xml = simplexml_load_string($this->client->GetExchangeRates(['startDate' => $date, 'endDate' => $date, 'currencyNames' => $code])->GetExchangeRatesResult)->Day;
+
+                return [
+                    'rate' => (float) $xml->Rate,
+                    'unit' => (int) $xml->Rate->attributes()->unit,
+                    'curr' => (string) $xml->Rate->attributes()->curr,
+                ];
+            }
+        );
     }
 
     /**
      * Will return with the current exchange rate
-     *
-     * @param $code
-     * @param null|Carbon $date
-     * @return null|\SzuniSoft\Mnb\Model\Currency
      */
-    public function currentExchangeRate($code, &$date = null)
+    public function currentExchangeRates(): array
     {
-        $result = $this->client->currentExchangeRate($code, $date);
-        $this->normalizeDate($date);
-        return $result;
+        return $this->cache->remember(config('mnb-exchange.cache.key').'.current', config('mnb-exchange.cache.minutes'), function () {
+            $current = [];
+            foreach (simplexml_load_string($this->client->GetCurrentExchangeRates()->GetCurrentExchangeRatesResult)->Day->Rate as $rate) {
+                $current[(string) $rate->attributes()->curr] = [
+                    'rate' => (float) $rate,
+                    'unit' => (int) $rate->attributes()->unit,
+                ];
+            }
+            return $current;
+        });
     }
 
-    /**
-     * @param null $date
-     * @return array|\SzuniSoft\Mnb\Model\Currency[]
-     */
-    public function currentExchangeRates(&$date = null)
+    public function lastOpeningDate(): string
     {
-        $result = $this->client->currentExchangeRates($date);
-        $this->normalizeDate($date);
-        return $result;
+        return (string) simplexml_load_string($this->client->GetDateInterval()->GetDateIntervalResult)->DateInterval->attributes()->enddate;
     }
 
+    public function firstOpeningDate(): string
+    {
+        return (string) simplexml_load_string($this->client->GetDateInterval()->GetDateIntervalResult)->DateInterval->attributes()->startdate;
+    }
 }
