@@ -5,7 +5,7 @@ return [
     /*
      * Wsdl file location.
      * */
-    'wsdl' => env('MNB_SOAP_WSDL', 'http://www.mnb.hu/arfolyamok.asmx?wsdl'),
+    'wsdl' => env('MNB_SOAP_WSDL', 'https://mnb.hu/arfolyamok.asmx?wsdl'),
 
     'cache' => [
 
@@ -19,6 +19,11 @@ return [
          * Default: 24hrs (1440)
          * */
         'minutes' => env('MNB_CACHE_MINUTES', 1440),
-    ]
+
+        /*
+         * Desired cache key prefix.
+         * */
+        'key' => env('MNB_CACHE_KEY', 'mnb'),
+    ],
 
 ];
