@@ -1,15 +1,11 @@
 <?php
 
 
-namespace SzuniSoft\Mnb\Laravel;
+namespace Csaba215\Mnb\Laravel;
 
 
 use Illuminate\Support\ServiceProvider;
 
-/**
- * Class MnbServiceProvider
- * @package SzuniSoft\Mnb\Laravel
- */
 class MnbServiceProvider extends ServiceProvider {
 
     /**
@@ -18,22 +14,6 @@ class MnbServiceProvider extends ServiceProvider {
     public function register()
     {
         $this->mergeConfigFrom(__DIR__ . '/../config/config.php', 'mnb-exchange');
-
-        $this->registerClient();
-    }
-
-    /**
-     *
-     */
-    protected function registerClient()
-    {
-        $this->app->singleton(Client::class, function ($app) {
-            $config = $app['config']['mnb-exchange'];
-            $cacheStore = $app['cache']->store($config['cache']['store']);
-            $mnbClient = new \SzuniSoft\Mnb\Client($config['wsdl']);
-            return new Client($mnbClient, $cacheStore,$config['cache']['minutes']);
-        });
-
         $this->app->alias(Client::class, 'mnb.client');
     }
 
